requirements.txt
Flask==3.0.2
spotipy==2.23.0
Flask-Login==0.6.2
python-dotenv==1.0.0

